Update patch set 2
Patch Set 2:

(1 comment)

These are the script for swift provisioning and not going to change. We had done sit down review with Michael and Ram. Only outstanding issue is to resolve missing container in CI registry and handling of script failures that we will handle in separate bug fix

Patch-set: 2