Update patch set 2
Patch Set 2: -Code-Review

> can we fail deployment if both tsn and compute roles are set for
 > node?

For an all-in-one config with TSN role, this is required because kolla playbooks fail if there is no nova-compute running in the cluster. So we let it come up and then disable the service.

Patch-set: 2
Reviewer: Gerrit User 330 <330@c72cd27d-1743-4a55-84a4-cd13b8400c52>
Label: Code-Review=0