Fix for image upload in Swift HA cluster
The default swift object container creation should not be done on all the target
nodes. It is now done on the first storage node and then the created
container/object files are copied into the other storage nodes.

This is a follow-up to Change-Id: Ie5b5ebb5f776f5c13f9775432043364752fd4efd

Change-Id: I7777a83a2182ea365df47e4ee6cf91a4bdfca6e0
Closes-Bug: #1782830