Update patch set 1
Patch Set 1:

Current use case is for openstack.
if k8s is deployed and needed by appformix, it will import k8s specific configs.
This info is openstack specific, not present explicitly in instances.yaml. Its not good practice to derive/compute the result in 2 separate modules. Its better for maintainability/correctness to keep that in single module and let rest of the modules use it. Hence exporting here.

Patch-set: 1