Certificates for TLS support for haproxy
When tls is enabled for haproxy by setting kolla_enable_tls_external knob, the
server certificate needs to be set using kolla_external_fqdn_cert (to a file
that exists on the ansible controller node).

If it is required to generate the certificates (self-signed), then set the
generate_self_signed_certs to 'yes'.

Change-Id: I5bb59be4c22fb34532dc3d49c3eb96de57be7fe2
Partial-Bug: #1807296