Update patch set 4
Patch Set 4:

> why changes are needed in vanilla kolla-ansible code?
 > How it can be achieved by user in OpenStack version of
 > kolla-ansible?

We want to generate certs if user does not provide his own certs without changes to current workflow of 3 playbooks (configure_instances, install_openstack and install_contrail). In that case, we need to introduce a new variable to conditionally run the generate certs playbook. So the conditional variable needs to be inside the certificates.yml playbook.

Patch-set: 4